Match a few OS files

+++ src/kyoshin/appgame/code_802AEB74.hpp
@@ -4,28 +4,28 @@
 
 //namespace cf {
 extern "C" {
-extern u32** lbl_80665708;
-extern u32** lbl_8066570C;
-extern u32** lbl_80665710;
-extern u32** lbl_80665714;
-extern u32** lbl_80665718;
-extern u32** lbl_8066571C;
-extern u32** lbl_80665720;
-extern u32** lbl_80665724;
-extern u32** lbl_80665728;
-extern u32** lbl_8066572C;
+extern u16** lbl_80665708;
+extern u16** lbl_8066570C;
+extern u16** lbl_80665710;
+extern u16** lbl_80665714;
+extern u16** lbl_80665718;
+extern u16** lbl_8066571C;
+extern u16** lbl_80665720;
+extern u16** lbl_80665724;
+extern u16** lbl_80665728;
+extern u16** lbl_8066572C;
 }
 
-u32** func_802AEB74();
-u32** func_802AEB7C();
-u32** func_802AEB84();
-u32** func_802AEB8C();
-u32** func_802AEB94();
-u32** func_802AEB9C();
-u32** func_802AEBA4();
-u32** func_802AEBAC();
-u32** func_802AEBB4();
-u32** func_802AEBBC();
+u16** func_802AEB74();
+u16** func_802AEB7C();
+u16** func_802AEB84();
+u16** func_802AEB8C();
+u16** func_802AEB94();
+u16** func_802AEB9C();
+u16** func_802AEBA4();
+u16** func_802AEBAC();
+u16** func_802AEBB4();
+u16** func_802AEBBC();
 
 
 //}
